+++ .ci/deploy/opendut-theo/src/commands/edgar/mod.rs
@@ -3,7 +3,7 @@ use std::time::Duration;
 
 use anyhow::{anyhow, Error};
 use tracing::debug;
-use opendut_edgar_kernel_modules::{default_builtin_module_dir, default_module_file, required_kernel_modules};
+use opendut_edgar_kernel_modules::{default_builtin_module_dir, default_module_file, required_can_kernel_modules};
 use status::EdgarDeploymentStatus;
 use crate::core::TheoError;
 use crate::core::docker::command::DockerCommand;
@@ -58,7 +58,7 @@ impl TestEdgarCli {
 }
 
 fn load_edgar_kernel_modules() -> Result<(), Error> {
-    for kernel_module in required_kernel_modules() {
+    for kernel_module in required_can_kernel_modules() {
         if !kernel_module.is_loaded(&default_module_file(), &default_builtin_module_dir())? {
             kernel_module.load()?;
         }