+++ opendut-edgar/src/setup/constants.rs
@@ -88,4 +88,5 @@ pub mod rperf {
     }
 }
 
-pub const REQUIRED_COMMAND_LINE_PROGRAMS: [(&str, &str); 3] = [("systemctl", "--version"), ("cannelloni", "-h"), ("cangw", "-s")];
+pub const REQUIRED_COMMAND_LINE_PROGRAMS: [(&str, &str); 1] = [("systemctl", "--version")];
+pub const REQUIRED_COMMAND_LINE_PROGRAMS_CAN: [(&str, &str); 2] = [("cannelloni", "-h"), ("cangw", "-s")];