+++ opendut-edgar/src/setup/tasks/can/mod.rs
@@ -1,0 +1,2 @@
+pub mod load_kernel_modules;
+pub mod create_kernel_module_load_rule;